split config into multiple files
