+++ src/main/java/de/melanx/skyblockbuilder/commands/invitation/DeclineCommand.java
@@ -4,7 +4,7 @@ import com.mojang.brigadier.arguments.StringArgumentType;
 import com.mojang.brigadier.builder.ArgumentBuilder;
 import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import de.melanx.skyblockbuilder.commands.Suggestions;
-import de.melanx.skyblockbuilder.config.ConfigHandler;
+import de.melanx.skyblockbuilder.config.common.PermissionsConfig;
 import de.melanx.skyblockbuilder.data.SkyblockSavedData;
 import de.melanx.skyblockbuilder.data.Team;
 import de.melanx.skyblockbuilder.events.SkyblockHooks;
@@ -47,7 +47,7 @@ public class DeclineCommand {
                 source.sendSuccess(() -> Component.translatable("skyblockbuilder.command.denied.decline_invitations").withStyle(ChatFormatting.RED), false);
                 return 0;
             case DEFAULT:
-                if (!ConfigHandler.Utility.selfManage && !source.hasPermission(2)) {
+                if (!PermissionsConfig.selfManage && !source.hasPermission(2)) {
                     source.sendSuccess(() -> Component.translatable("skyblockbuilder.command.disabled.decline_invitations").withStyle(ChatFormatting.RED), false);
                     return 0;
                 }
