+++ src/main/java/de/melanx/skyblockbuilder/config/common/PermissionsConfig.java
@@ -1,0 +1,48 @@
+package de.melanx.skyblockbuilder.config.common;
+
+import org.moddingx.libx.annotation.config.RegisterConfig;
+import org.moddingx.libx.config.Config;
+import org.moddingx.libx.util.data.ResourceList;
+
+@RegisterConfig("permissions")
+public class PermissionsConfig {
+
+    @Config("Should players be able to leave their team or invite others? [default: true]")
+    public static boolean selfManage = true;
+
+    @Config("Should players be able to create their own team? [default: false]")
+    public static boolean createOwnTeam = false;
+
+    public static class Teleports {
+
+        @Config("Should players be able to teleport to spawn? [default: true]")
+        public static boolean spawn = true;
+
+        @Config("Cooldown in ticks for teleporting to spawn. [default: 3600 = 3min]")
+        public static int spawnCooldown = 3600;
+
+        @Config("Should players be able to visit other island? [default: true]")
+        public static boolean allowVisits = true;
+
+        @Config("Should players be able to teleport to their home island? [default: true]")
+        public static boolean home = true;
+
+        @Config("Cooldown in ticks for teleporting back home. [default: 3600 = 3min]")
+        public static int homeCooldown = 3600;
+
+        @Config("Should players be able to teleport to another dimension? [default: true]")
+        public static boolean crossDimensionTeleportation = true;
+
+        @Config("Dimensions in this list are not allowed for executing teleportation commands. Inverted behaviour if you set \"allow_list\" to true.")
+        public static ResourceList teleportationDimensions = ResourceList.DENY_LIST;
+    }
+
+    public static class Spawns {
+
+        @Config("The range from island center for possible spawns to add. [default: 50]")
+        public static int range = 50;
+
+        @Config("Should players be able to modify their spawn positions? [default: true]")
+        public static boolean modifySpawns = true;
+    }
+}
